+++ mobile/android/android-components/components/feature/tabs/src/main/java/mozilla/components/feature/tabs/toolbar/TabsToolbarFeature.kt
@@ -4,25 +4,22 @@
 
 package mozilla.components.feature.tabs.toolbar
 
-import android.content.Context
+import mozilla.components.browser.session.SessionManager
 import mozilla.components.concept.toolbar.Toolbar
-import mozilla.components.feature.tabs.R
 
 /**
  * Feature implementation for connecting a tabs tray implementation with a toolbar implementation.
  */
 
 class TabsToolbarFeature(
-    context: Context,
     toolbar: Toolbar,
+    sessionManager: SessionManager,
     showTabs: () -> Unit
 ) {
     init {
-        val tabsAction = Toolbar.ActionButton(
-            mozilla.components.ui.icons.R.drawable.mozac_ic_tab,
-            context.getString(R.string.mozac_feature_tabs_toolbar_tabs_button)) {
-            showTabs.invoke()
-        }
+        val tabsAction = TabCounterToolbarButton(
+            sessionManager,
+            showTabs)
 
         toolbar.addBrowserAction(tabsAction)
     }