+++ mobile/android/android-components/samples/browser/src/main/java/org/mozilla/samples/browser/BrowserFragment.kt
@@ -58,7 +58,7 @@ class BrowserFragment : Fragment(), BackHandler, DownloadDialogListener {
                 components.defaultSearchUseCase,
                 sessionId)
 
-        tabsToolbarFeature = TabsToolbarFeature(requireContext(), toolbar, ::showTabs)
+        tabsToolbarFeature = TabsToolbarFeature(toolbar, components.sessionManager, ::showTabs)
 
         downloadsFeature = DownloadsFeature(
             requireContext(),