+++ mobile/android/android-components/docs/changelog.md
@@ -118,6 +118,8 @@ permalink: /changelog/
   ```
   * Showing the crash prompt won't play the default activity animation anymore.
   * Added a new sample app `samples-crash` to show and test crash reporter integration.
+* **feature-tabs**:
+  * `TabsToolbarFeature` is now adding a `TabCounter` from the `ui-tabcounter` component to the toolbar.
 * **lib-jexl**
   * New component for evaluating Javascript Expression Language (JEXL) expressions. This implementation is based on [Mozjexl](https://github.com/mozilla/mozjexl) used at Mozilla, specifically as a part of SHIELD and Normandy. In a future version of Fretboard JEXL will allow more complex rules for experiments. For more see [documentation](https://github.com/mozilla-mobile/android-components/blob/master/components/lib/jexl/README.md).
 * **service-telemetry**